refactor: restructured utilities
