+++ src/content-scripts/profileScreen.ts
@@ -1,7 +1,7 @@
 import { getGithubUsername } from "../utils/matchers";
 import { getOpenSaucedUser } from "../utils/fetchOpenSaucedApiData";
-import injectViewOnOpenSauced from "../utils/viewOnOpenSauced";
-import injectInviteToOpenSauced from "../utils/inviteToOpenSauced";
+import injectViewOnOpenSauced from "../utils/dom-utils/viewOnOpenSauced";
+import injectInviteToOpenSauced from "../utils/dom-utils/inviteToOpenSauced";
 
 const username = getGithubUsername(window.location.href);
 if (username != null) {
